+++ src/main/groovy/rocks/metaldetector/butler/config/security/SecurityConfig.groovy
@@ -1,13 +1,10 @@
-package rocks.metaldetector.butler.config
+package rocks.metaldetector.butler.config.security
 
 import org.springframework.context.annotation.Configuration
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.builders.WebSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter
-import org.springframework.security.web.util.matcher.RequestMatcher
-
-import javax.servlet.http.HttpServletRequest
 
 import static org.springframework.http.HttpMethod.GET
 import static org.springframework.http.HttpMethod.POST
@@ -27,13 +24,5 @@ class SecurityConfig extends WebSecurityConfigurerAdapter {
   @Override
   void configure(WebSecurity web) throws Exception {
     web.ignoring().antMatchers(POST, "/**") // temporary until implementation of https://trello.com/c/x1MpE505
-  }
-}
-
-class XForwardedProtoMatcher implements RequestMatcher {
-
-  @Override
-  boolean matches(HttpServletRequest request) {
-    request.getHeader("X-Forwarded-Proto") != null
   }
 }
