+++ src/main/groovy/rocks/metaldetector/butler/config/swagger/SwaggerSpecConfig.groovy
@@ -1,4 +1,4 @@
-package rocks.metaldetector.butler.config
+package rocks.metaldetector.butler.config.swagger
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
