+++ src/test/groovy/rocks/metaldetector/butler/web/rest/ReleasesRestControllerIT.groovy
@@ -11,7 +11,7 @@ import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
 import org.springframework.test.web.servlet.MockMvc
-import rocks.metaldetector.butler.config.Endpoints
+import rocks.metaldetector.butler.config.constants.Endpoints
 import rocks.metaldetector.butler.model.TimeRange
 import rocks.metaldetector.butler.service.ReleaseServiceImpl
 import rocks.metaldetector.butler.testutil.WithIntegrationTestProfile
